+++ server/src/modules/auth/github.service.ts
@@ -1,12 +1,10 @@
 import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { HttpService } from "@nestjs/axios";
-
-const { OAUTH_GITHUB_CLIENT_ID, OAUTH_GITHUB_CLIENT_SECRET, OAUTH_GITHUB_ACCESS_TOKEN_API, OAUTH_GITHUB_EMAIL_API } =
-  process.env;
+import { ConfigService } from "@nestjs/config";
 
 @Injectable()
 export class GithubService {
-  constructor(private readonly httpService: HttpService) {}
+  constructor(private readonly httpService: HttpService, private readonly configService: ConfigService) {}
 
   async getGithubEmail(code: string) {
     const accessToken = await this.fetchGithubAccessToken(code);
@@ -22,14 +20,14 @@ export class GithubService {
     };
 
     const accesstokenAPIData = {
-      client_id: OAUTH_GITHUB_CLIENT_ID,
-      client_secret: OAUTH_GITHUB_CLIENT_SECRET,
+      client_id: this.configService.get<string>("OAUTH_GITHUB_CLIENT_ID"),
+      client_secret: this.configService.get<string>("OAUTH_GITHUB_CLIENT_SECRET"),
       code,
     };
 
     try {
       const res = await this.httpService
-        .post(OAUTH_GITHUB_ACCESS_TOKEN_API, accesstokenAPIData, accesstokenAPIConfig)
+        .post(this.configService.get<string>("OAUTH_GITHUB_ACCESS_TOKEN_API"), accesstokenAPIData, accesstokenAPIConfig)
         .toPromise();
       const { access_token: accessToken } = res.data;
       return accessToken;
@@ -47,7 +45,9 @@ export class GithubService {
     };
 
     try {
-      const res = await this.httpService.get(OAUTH_GITHUB_EMAIL_API, githubUserAPIConfig).toPromise();
+      const res = await this.httpService
+        .get(this.configService.get<string>("OAUTH_GITHUB_EMAIL_API"), githubUserAPIConfig)
+        .toPromise();
       const email = res.data[0].email;
       return email;
     } catch (err) {
