+++ server/src/modules/auth/jwt/jwt-access-token.strategy.ts
@@ -1,9 +1,10 @@
 import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { PassportStrategy } from "@nestjs/passport";
 import { ExtractJwt, Strategy } from "passport-jwt";
-import { PrismaService } from "src/prisma/prisma.service";
+import { PrismaService } from "src/modules/prisma/prisma.service";
 import { User } from "@prisma/client";
 import { Request } from "express";
+import { ConfigService } from "@nestjs/config";
 
 interface JwtPayload {
   userId: number;
@@ -11,9 +12,9 @@ interface JwtPayload {
 
 @Injectable()
 export class AccessTokenStrategy extends PassportStrategy(Strategy, "jwt-access-token") {
-  constructor(private prisma: PrismaService) {
+  constructor(private config: ConfigService, private prisma: PrismaService) {
     super({
-      secretOrKey: process.env.JWT_SECRET_KEY,
+      secretOrKey: config.get<string>("JWT_SECRET_KEY"),
       jwtFromRequest: ExtractJwt.fromExtractors([
         (req: Request) => {
           return req.cookies["jwt-access-token"];
