+++ server/src/modules/auth/jwt/jwt-refresh-token.strategy.ts
@@ -1,8 +1,9 @@
 import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { PassportStrategy } from "@nestjs/passport";
 import { ExtractJwt, Strategy } from "passport-jwt";
-import { PrismaService } from "src/prisma/prisma.service";
+import { PrismaService } from "src/modules/prisma/prisma.service";
 import { Request } from "express";
+import { ConfigService } from "@nestjs/config";
 
 interface JwtPayload {
   userId: number;
@@ -10,9 +11,9 @@ interface JwtPayload {
 
 @Injectable()
 export class RefreshTokenStrategy extends PassportStrategy(Strategy, "jwt-refresh-token") {
-  constructor(private prisma: PrismaService) {
+  constructor(private config: ConfigService, private prisma: PrismaService) {
     super({
-      secretOrKey: process.env.JWT_SECRET_KEY,
+      secretOrKey: config.get<string>("JWT_SECRET_KEY"),
       jwtFromRequest: ExtractJwt.fromExtractors([
         (req: Request) => {
           return req.cookies["jwt-refresh-token"];
