+++ server/src/modules/game/game.module.ts
@@ -3,7 +3,7 @@ import { Module } from "@nestjs/common";
 import { GameController } from "./game.controller";
 
 import { GameService } from "./game.service";
-import { PrismaModule } from "src/prisma/prisma.module";
+import { PrismaModule } from "src/modules/prisma/prisma.module";
 
 @Module({
   imports: [PrismaModule],
